requirements.txt
uvicorn~=0.17.6
quart~=0.16.3
aiohttp~=3.8.1
python-dotenv~=0.19.2
pydantic~=1.9.0
python-dateutil~=2.8.2